requirements.txt
# FastAPI and Server Dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
python-decouple==3.8
argon2-cffi==23.1.0

# Database
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# External APIs and Services
openai==1.12.0
google-search-results==2.4.2  # SerpAPI
selenium==4.18.0
beautifulsoup4==4.12.3
requests==2.31.0

# Data Processing and Analysis
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3

# Testing
pytest==8.0.2
httpx==0.27.0
pytest-asyncio==0.23.5